refactor: split god-modules into domain layout (#52)
## Summary

Restructures the codebase into clean domain modules so that as it grows,
it stays clear where each thing lives. No behavior changes — pure
reorganization, with build & smoke test green at every step.

- `update/` — version polling + release-notes state, exposed as
`UpdateStatus`
- `ha/` — connection lifecycle + token storage in one place (`HaState`,
`ha::token`)
- `app/` — split from a 908-line god module into `snapdash.rs` (state +
message core), `lifecycle.rs` (iced runtime wiring), `window.rs`
(multi-window types)
- `ui/update_view.rs` — small UI bridge that maps domain `UpdateState`
to a styled icon, keeps the update domain free of `Icon`/`Palette` deps

`lib.rs` is unchanged: `Snapdash::boot/update/view/subscription/style`
still resolve via multi-impl blocks across the new files.

## Commits (one logical step each)

1. `cb301e10` — extract `UpdateStatus` into `update/` module
2. `3e15f648` — move `secrets.rs` into `ha/token.rs`, rename fns to
`set/get/delete`
3. `04a67d3a` — consolidate four HA fields on `Snapdash` into
`ha::HaState`
4. `9d49d076` — convert `app.rs` to `app/` directory (`app/snapdash.rs`
+ thin `mod.rs`)
5. `6048c635` — extract window types + `find_window_id` to
`app/window.rs`
6. `53675236` — move `boot` and `subscription` to `app/lifecycle.rs`
7. `d6f73450` — annotate the new module layout with module docstrings +
targeted comments

## Notes / non-goals

- `Message::OpenReleaseNotes` and `WindowKind::ReleaseNotes`
deliberately stay in app — they are window-management concerns, not
update-domain.
- `Message` enum is not split per-domain; that would be a substantially
bigger change and only really pays off if/when this becomes a Cargo
workspace.
- `entity_window.rs` still uses Unicode emoji (`✕`, `⚙`) instead of
`Icon::*` — out of scope, follow-up.
- `app/snapdash.rs` is still 775 lines (the big `update()` match). A
natural next step is splitting handlers by domain
(`app/handlers/{ha,window,update,config}.rs`).



---------
