+++ src/ui/mod.rs
@@ -7,3 +7,4 @@ pub mod platform;
 pub mod release_notes;
 pub mod settings;
 pub mod theme;
+pub mod update_view;
